packages/web_fetcher/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-fetcher"
version = "1.0.0"
description = "Production-ready PDF fetcher with DOI resolution, YAML configuration, and Cloudflare handling"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Henrik Kragh Sørensen"}
]
keywords = ["web", "fetcher", "scraping", "selenium", "caching", "http", "pdf", "doi", "academic"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "urllib3>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
selenium = [
    "selenium>=4.15.0",
]
progress = [
    "tqdm>=4.65.0",
]
full = [
    "selenium>=4.15.0",
    "tqdm>=4.65.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]

[tool.setuptools]
packages = ["web_fetcher"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
